requirements.txt
absl-py==2.0.0
asttokens==2.4.1
astunparse==1.6.3
blinker==1.7.0
breadability==0.1.20
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
clang==5.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.28
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
flatbuffers==1.12
fonttools==4.44.0
fsspec==2023.10.0
gast==0.4.0
gensim-sum-ext==0.1.2
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.1.0
idna==3.4
importlib-metadata==6.8.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==1.4.4
parso==0.8.3
Pillow==10.1.0
plotly==5.18.0
prompt-toolkit==3.0.40
protobuf==3.9.2
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycorenlp==0.3.0
pycountry==22.3.5
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.15.0
stack-data==0.6.3
sumy==0.10.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
traitlets==5.13.0
typing-extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
wrapt==1.12.1
zipp==3.17.0
ultralytics